Add GitHub Actions example, note no token needed
The first line of the `codecov-action` README notes:

> The latest release of this Action adds support for tokenless uploads
> from GitHub Actions!

With Travis CI changing its open source offerings, many projects will
likely be switching to GitHub Actions. Add a brief example, and a link
back to the repository's README.

Also leave a note in the FAQ that tokens are not required if using the
GitHub action.